// face-check-ui/src/app/services/models/page-response-related-user-in-company-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { RelatedUserInCompanyResponse } from '../models/related-user-in-company-response';
export interface PageResponseRelatedUserInCompanyResponse {
  content?: Array<RelatedUserInCompanyResponse>;
  first?: boolean;
  last?: boolean;
  number?: number;
  size?: number;
  totalElement?: number;
  totalPages?: number;
}